[JSC] Fix integer overflow on armv7 in CompleteSubspace::tryAllocateSlow https://bugs.webkit.org/show_bug.cgi?id=289157
Reviewed by Yusuke Suzuki.

Avoids integer overflow when multiplying by ramSize.

* Source/JavaScriptCore/heap/CompleteSubspace.cpp:
(JSC::CompleteSubspace::tryAllocateSlow):

Canonical link: https://commits.webkit.org/291887@main

+++ Source/JavaScriptCore/heap/CompleteSubspace.cpp
@@ -130,7 +130,7 @@ void* CompleteSubspace::tryAllocateSlow(VM& vm, size_t size, GCDeferralContext* 
     
     vm.heap.collectIfNecessaryOrDefer(deferralContext);
     if (UNLIKELY(Options::maxHeapSizeAsRAMSizeMultiple())) {
-        if (vm.heap.capacity() > Options::maxHeapSizeAsRAMSizeMultiple() * WTF::ramSize())
+        if (vm.heap.capacity() > static_cast<uint64_t>(Options::maxHeapSizeAsRAMSizeMultiple()) * static_cast<uint64_t>(WTF::ramSize()))
             return nullptr;
     }
